3d cloned a quadratic demo
