+++ compact.cirru
@@ -51,7 +51,7 @@
             def skip-rendering? $ = "\"true" (get-env "\"skip" "\"false")
         |tabs $ %{} :CodeEntry (:doc |)
           :code $ quote
-            def tabs $ [] (:: :fireworks |Fireworks :dark) (:: :lorenz |Lorenz :dark) (:: :aizawa |Aizawa :dark) (:: :fourwing "|Four Wing" :dark) (:: :fractal |Fractal :dark) (:: :collision |Collision :dark) (:: :bounce |Bounce :dark) (:: :feday |FEDAY :dark) (:: :bifurcation "\"Bifurcation" :dark) (:: :ball-spin "\"Ball Spin" :dark) (:: :lifegame "\"Lifegame" :dark) (:: :lifegame-trail "\"Lifegame Trail" :dark) (:: :bounce-trail "|Bounce Trail" :dark) (:: :orbit-spark "|Orbit Spark" :dark) (:: :chen |Chen :dark) (:: :sprott |Sprott :dark) (:: :lorenz83 |Lorenz83 :dark) (:: :orbits |Orbits :dark) (:: :lamps |Lamps :dark) (:: :debug-grid "|Debug Grid" :dark) (:: :den-tsucs "\"Den Tsucs" :dark) (:: :bouali "\"Bouali" :dark) (:: :orbits2 "\"Orbits 2" :dark) (:: :halvorsen "\"Halvorsen" :dark) (:: :clifford "\"Clifford" :dark) (:: :dequanli "\"Dequan Li" :dark) (:: :dadras "\"Dadras" :dark) (:: :burke-shaw "\"Burke Shaw" :dark)
+            def tabs $ [] (:: :fireworks |Fireworks :dark) (:: :lorenz |Lorenz :dark) (:: :aizawa |Aizawa :dark) (:: :fourwing "|Four Wing" :dark) (:: :fractal |Fractal :dark) (:: :collision |Collision :dark) (:: :bounce |Bounce :dark) (:: :feday |FEDAY :dark) (:: :bifurcation "\"Bifurcation" :dark) (:: :ball-spin "\"Ball Spin" :dark) (:: :lifegame "\"Lifegame" :dark) (:: :lifegame-trail "\"Lifegame Trail" :dark) (:: :bounce-trail "|Bounce Trail" :dark) (:: :orbit-spark "|Orbit Spark" :dark) (:: :chen |Chen :dark) (:: :sprott |Sprott :dark) (:: :lorenz83 |Lorenz83 :dark) (:: :orbits |Orbits :dark) (:: :lamps |Lamps :dark) (:: :debug-grid "|Debug Grid" :dark) (:: :den-tsucs "\"Den Tsucs" :dark) (:: :bouali "\"Bouali" :dark) (:: :orbits2 "\"Orbits 2" :dark) (:: :halvorsen "\"Halvorsen" :dark) (:: :clifford "\"Clifford" :dark) (:: :dequanli "\"Dequan Li" :dark) (:: :dadras "\"Dadras" :dark) (:: :burke-shaw "\"Burke Shaw" :dark) (:: :quadratic "\"Quadratic" :dark)
         |threshold $ %{} :CodeEntry (:doc |)
           :code $ quote
             def threshold $ js/parseFloat
@@ -151,6 +151,7 @@
                 :dequanli $ dequanli/loadRenderer canvas
                 :dadras $ dadras/loadRenderer canvas
                 :burke-shaw $ burke-shaw/loadRenderer canvas
+                :quadratic $ quadratic/loadRenderer canvas
         |reload! $ %{} :CodeEntry (:doc |)
           :code $ quote
             defn reload! () (hint-fn async)
@@ -224,6 +225,7 @@
             "\"../src/apps/attractor/dequanli" :as dequanli
             "\"../src/apps/attractor/dadras" :as dadras
             "\"../src/apps/attractor/burke-shaw" :as burke-shaw
+            "\"../src/apps/quadratic" :as quadratic
             "\"../src/index" :refer $ setupInitials
             "\"../src/config" :as js-config
             "\"../src/index" :refer $ listenShaderError
